+++ drivers/versati/driver.ts
@@ -24,6 +24,7 @@ interface FlowDevice {
   flowSetWeatherDependent(enabled: unknown): Promise<void>;
   flowSetDisinfect(enabled: unknown): Promise<void>;
   flowSetCurveOutdoorTemperature(temperature: unknown): Promise<void>;
+  flowSetCopElectricalInput(power: unknown): Promise<void>;
   flowPauseWeatherCurve(minutes: unknown): Promise<void>;
   flowResumeWeatherCurve(): Promise<void>;
   flowSetWeatherCurveBoost(offset: unknown, minutes: unknown): Promise<void>;
@@ -36,6 +37,8 @@ interface FlowDevice {
   flowCurveSkippedReasonIs(reason: unknown): boolean;
   flowCurveWriteAllowed(): boolean;
   flowHasPollHistory(): boolean;
+  flowCopBelow(cop: unknown): boolean;
+  flowCopStatusIs(status: unknown): boolean;
 }
 
 class GreeVersatiDriver extends Homey.Driver {
@@ -95,6 +98,10 @@ class GreeVersatiDriver extends Homey.Driver {
 
     this.homey.flow.getActionCard('set_curve_outdoor_temperature').registerRunListener(async (args) => {
       await flowDevice(args).flowSetCurveOutdoorTemperature(numberValue(args.temperature));
+    });
+
+    this.homey.flow.getActionCard('set_cop_electrical_input').registerRunListener(async (args) => {
+      await flowDevice(args).flowSetCopElectricalInput(numberValue(args.power));
     });
 
     this.homey.flow.getActionCard('pause_weather_curve').registerRunListener(async (args) => {
@@ -159,6 +166,14 @@ class GreeVersatiDriver extends Homey.Driver {
 
     this.homey.flow.getConditionCard('has_poll_history').registerRunListener((args) => {
       return flowDevice(args).flowHasPollHistory();
+    });
+
+    this.homey.flow.getConditionCard('cop_below').registerRunListener((args) => {
+      return flowDevice(args).flowCopBelow(numberValue(args.cop));
+    });
+
+    this.homey.flow.getConditionCard('cop_status_is').registerRunListener((args) => {
+      return flowDevice(args).flowCopStatusIs(dropdownValue(args.status));
     });
   }
 }